+++ examples/training.py
@@ -34,12 +34,8 @@ train_ds = Dataset(train_records, train_tfms)
 valid_ds = Dataset(valid_records, valid_tfms)
 
 # Create both training and validation dataloaders
-train_dl = faster_rcnn.train_dataloader(
-    train_ds, batch_size=16, num_workers=4, shuffle=True
-)
-valid_dl = faster_rcnn.valid_dataloader(
-    valid_ds, batch_size=16, num_workers=4, shuffle=False
-)
+train_dl = faster_rcnn.train_dl(train_ds, batch_size=16, num_workers=4, shuffle=True)
+valid_dl = faster_rcnn.valid_dl(valid_ds, batch_size=16, num_workers=4, shuffle=False)
 
 # Create model
 model = faster_rcnn.model(num_classes=len(class_map))