+++ mantisshrimp/datasets/birds/__init__.py
@@ -1,0 +1,2 @@
+from mantisshrimp.datasets.birds.data import *
+from mantisshrimp.datasets.birds.parsers import *