+++ mantisshrimp/models/efficientdet/dataloaders.py
@@ -2,16 +2,16 @@ __all__ = [
     "build_train_batch",
     "build_valid_batch",
     "build_infer_batch",
-    "train_dataloader",
-    "valid_dataloader",
-    "infer_dataloader",
+    "train_dl",
+    "valid_dl",
+    "infer_dl",
 ]
 
 from mantisshrimp.imports import *
 from mantisshrimp.models.utils import *
 
 
-def train_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
+def train_dl(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
     """ A `DataLoader` with a custom `collate_fn` that batches items as required for training the model.
 
     # Arguments
@@ -23,7 +23,7 @@ def train_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoade
     # Returns
         A Pytorch `DataLoader`.
     """
-    return transform_dataloader(
+    return transform_dl(
         dataset=dataset,
         build_batch=build_train_batch,
         batch_tfms=batch_tfms,
@@ -31,7 +31,7 @@ def train_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoade
     )
 
 
-def valid_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
+def valid_dl(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
     """ A `DataLoader` with a custom `collate_fn` that batches items as required for validating the model.
 
     # Arguments
@@ -43,7 +43,7 @@ def valid_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoade
     # Returns
         A Pytorch `DataLoader`.
     """
-    return transform_dataloader(
+    return transform_dl(
         dataset=dataset,
         build_batch=build_valid_batch,
         batch_tfms=batch_tfms,
@@ -51,7 +51,7 @@ def valid_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoade
     )
 
 
-def infer_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
+def infer_dl(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoader:
     """ A `DataLoader` with a custom `collate_fn` that batches items as required for inferring the model.
 
     # Arguments
@@ -63,7 +63,7 @@ def infer_dataloader(dataset, batch_tfms=None, **dataloader_kwargs) -> DataLoade
     # Returns
         A Pytorch `DataLoader`.
     """
-    return transform_dataloader(
+    return transform_dl(
         dataset=dataset,
         build_batch=build_infer_batch,
         batch_tfms=batch_tfms,