+++ mantisshrimp/utils/download_utils.py
@@ -1,4 +1,9 @@
-__all__ = ["download_url", "download_and_extract"]
+__all__ = [
+    "download_url",
+    "download_and_extract",
+    "download_gdrive",
+    "download_and_extract_gdrive",
+]
 
 from mantisshrimp.imports import *
 import requests
@@ -23,3 +28,20 @@ def download_and_extract(
 
     download_url(url=str(url), save_path=str(save_path), chunk_size=chunk_size)
     shutil.unpack_archive(filename=str(save_path), extract_dir=str(save_dir))
+
+
+def download_gdrive(url, save_path):
+    """Download from gdrive, passing virus scan for big files.
+    """
+    import gdown
+
+    gdown.download(url=str(url), output=str(save_path), quiet=False)
+
+
+def download_and_extract_gdrive(url, filename, extract_dir):
+    extract_dir = Path(extract_dir)
+    save_path = extract_dir / filename
+    extract_dir.mkdir(exist_ok=True, parents=True)
+
+    download_gdrive(url=url, save_path=save_path)
+    shutil.unpack_archive(filename=str(save_path), extract_dir=str(extract_dir))